# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "用户管理系统",
        "contact": {},
        "version": "0.1"
    },
    "paths": {
        "/api/question/admin/create": {
            "post": {
                "description": "add question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Question"
                ],
                "summary": "add question",
                "parameters": [
                    {
                        "description": "add question",
                        "name": "question",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modelquestion.AddQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "add successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "add failed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/question/admin/delete": {
            "get": {
                "description": "delete question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Question"
                ],
                "summary": "delete question",
                "parameters": [
                    {
                        "type": "string",
                        "description": "delete Question",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete  successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "delete failed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/question/admin/update": {
            "post": {
                "description": "update question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Question"
                ],
                "summary": "update question",
                "parameters": [
                    {
                        "description": "update question",
                        "name": "question",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modelquestion.UpdateQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "update successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "update failed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/question/get": {
            "get": {
                "description": "get question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Question"
                ],
                "summary": "get question",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get Question",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get question successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/modelquestion.ReturnQuestion"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "EditUserRequest failed\"\t\"get question failed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "get questions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Question"
                ],
                "summary": "get questions",
                "parameters": [
                    {
                        "description": "Get Questions",
                        "name": "question",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modelquestion.QueryQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get questions successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/modelquestion.ReturnQuestion"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "EditUserRequest failed\"\t\"get questions failed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "modelquestion.AddQuestionRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "answer": {
                    "description": "\"题目答案\"",
                    "type": "string"
                },
                "content": {
                    "description": "\"内容\"",
                    "type": "string"
                },
                "judge_case": {
                    "description": "\"判题用例json数组\"",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/modelquestion.JudgeCase"
                    }
                },
                "judge_config": {
                    "description": "\"判题配置json对象\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/modelquestion.JudgeConfig"
                        }
                    ]
                },
                "tags": {
                    "description": "\"标签列表json数组\"",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/modelquestion.Tags"
                    }
                },
                "title": {
                    "description": "\"标题\"",
                    "type": "string"
                },
                "user_id": {
                    "description": "\"创建用户id\"",
                    "type": "string"
                }
            }
        },
...
        "modelquestion.JudgeCase": {
            "type": "object",
            "properties": {
                "input": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                }
            }
        },
        "modelquestion.JudgeConfig": {
            "type": "object",
            "required": [
                "memory_limit",
                "stack_limit",
                "time_limit"
            ],
            "properties": {
                "memory_limit": {
                    "type": "integer"
                },
                "stack_limit": {
                    "type": "integer"
                },
                "time_limit": {
                    "type": "integer"
                }
            }
        },
...
        "modelquestion.QueryQuestionRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "\"内容\"",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "judge_case": {
                    "description": "\"判题用例json数组\"",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/modelquestion.JudgeCase"
                    }
                },
                "tags": {
                    "description": "\"标签列表json数组\"",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/modelquestion.Tags"
                    }
                },
                "title": {
                    "description": "\"标题\"",
                    "type": "string"
                },
                "user_id": {
                    "description": "\"创建用户id\"",
                    "type": "string"
                }
            }
        },
...
        "modelquestion.ReturnQuestion": {
            "type": "object",
            "properties": {
                "accept_num": {
                    "description": "\"题目通过数\"",
                    "type": "integer"
                },
                "content": {
                    "description": "\"内容\"",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "judge_config": {
                    "description": "\"判题配置json对象\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/modelquestion.JudgeConfig"
                        }
                    ]
                },
                "submit_num": {
                    "description": "\"题目提交数",
                    "type": "integer"
                },
                "tags": {
                    "description": "\"标签列表json数组\"",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/modelquestion.Tags"
                    }
                },
                "thum_num": {
                    "description": "\"点赞数\"",
                    "type": "integer"
                },
                "title": {
                    "description": "\"标题\"",
                    "type": "string"
                },
                "user_id": {
                    "description": "用户id",
                    "type": "string"
                }
            }
        },
        "modelquestion.Tags": {
            "type": "object",
            "properties": {
                "easy": {
                    "type": "string"
                },
                "hard": {
                    "type": "string"
                },
                "medium": {
                    "type": "string"
                }
            }
        },
        "modelquestion.UpdateQuestionRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "answer": {
                    "description": "\"题目答案\"",
                    "type": "string"
                },
                "content": {
                    "description": "\"内容\"",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "judge_case": {
                    "description": "\"判题用例json数组\"",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/modelquestion.JudgeCase"
                    }
                },
                "judge_config": {
                    "description": "\"判题配置json对象\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/modelquestion.JudgeConfig"
                        }
                    ]
                },
                "tags": {
                    "description": "\"标签列表json数组\"",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/modelquestion.Tags"
                    }
                },
                "title": {
                    "description": "\"标题\"",
                    "type": "string"
                }
            }
        },
...
        "utils.ApiResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}